Move to 64 bit netblock key

+++ inst/include/netblockr.h
@@ -1,4 +1,19 @@
-#include <stdint.h>
+// Copyright (C) 2017  Jon Meek
+//
+// This file is part of netblockr.
+//
+// netblockr is free software: you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 2 of the License, or
+// (at your option) any later version.
+//
+// netblockr is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with netblockr.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <arpa/inet.h>
 #include <iostream>
@@ -17,13 +32,12 @@ static const u_int maskval_ipv4[33] = { 0x00000000,
 					0xffffff80, 0xffffffc0, 0xffffffe0, 0xfffffff0,
 					0xfffffff8, 0xfffffffc, 0xfffffffe, 0xffffffff };
 
-
-// Data structure for the table of networks
-
+// The table of networks
 struct nbTable {
-  std::vector <u_int>       nb_base_as_uint;
+  std::vector <uint64_t>       nb_base_as_uint;
   std::vector <int>         nb_mask, nb_unique_masks;
   std::vector <std::string> nb_base_and_mask, nb_base_as_string, nb_description;
-  std::map <u_int, int>     nb_map;
+  std::map <uint64_t, int>     nb_map;
 };
+
